code: inherit custom errors from DetoxError
