+++ detox/local-cli/cli.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 global.DETOX_CLI = true;
 const yargs = require('yargs');
-const DetoxRuntimeError = require('../src/errors/DetoxRuntimeError');
+const DetoxError = require('../src/errors/DetoxError');
 const logger = require('../src/utils/logger').child({ __filename });
 
 yargs
@@ -25,7 +25,7 @@ yargs
   .wrap(yargs.terminalWidth() * 0.9)
   .fail(function(msg, err, program) {
     if (err) {
-      const lines = DetoxRuntimeError.format(err).split("\n");
+      const lines = DetoxError.format(err).split("\n");
       for (const line of lines) {
         logger.error(line);
       }
