+++ detox/runners/jest-circus/environment.js
@@ -5,7 +5,7 @@ const DetoxInitErrorListener = require('./listeners/DetoxInitErrorListener');
 const assertJestCircus26 = require('./utils/assertJestCircus26');
 const assertExistingContext = require('./utils/assertExistingContext');
 const wrapErrorWithNoopLifecycle = require('./utils/wrapErrorWithNoopLifecycle');
-const DetoxRuntimeError = require('../../src/errors/DetoxRuntimeError');
+const DetoxError = require('../../src/errors/DetoxError');
 const Timer = require('../../src/utils/Timer');
 
 const SYNC_CIRCUS_EVENTS = new Set([
@@ -149,7 +149,7 @@ class DetoxCircusEnvironment extends NodeEnvironment {
 
   /** @private */
   _logError(e) {
-    this._logger.error(DetoxRuntimeError.format(e));
+    this._logger.error(DetoxError.format(e));
   }
 
   /** @protected */
