+++ detox/src/errors/DetoxInternalError.js
@@ -1,15 +1,13 @@
-class DetoxInternalError extends Error {
+const DetoxError = require('./DetoxError');
+
+class DetoxInternalError extends DetoxError {
   constructor(message) {
-    super(message + '\n' + DetoxInternalError.reportIssue);
+    super(message + '\n' + DetoxError.reportIssue);
     this.name = 'DetoxInternalError';
   }
 
   static from(message) {
     return new DetoxInternalError(message).toString();
-  }
-
-  static get reportIssue() {
-    return 'Please report this issue on our GitHub tracker:\nhttps://github.com/wix/Detox/issues';
   }
 }
