+++ detox/src/server/DetoxConnection.js
@@ -1,7 +1,6 @@
 const _ = require('lodash');
 const { WebSocket } = require('ws');
 const DetoxRuntimeError = require('../errors/DetoxRuntimeError');
-const DetoxInternalError = require('../errors/DetoxInternalError');
 const AnonymousConnectionHandler = require('./handlers/AnonymousConnectionHandler');
 const logger = require('../utils/logger').child({ __filename });
 
@@ -62,7 +61,7 @@ class DetoxConnection {
         if (_.isError(action)) {
           throw new DetoxRuntimeError({
             message: 'The payload received is not a valid JSON.',
-            hint: DetoxInternalError.reportIssue,
+            hint: DetoxRuntimeError.reportIssue,
             debugInfo: data,
           });
         }
@@ -70,7 +69,7 @@ class DetoxConnection {
         if (!action.type) {
           throw new DetoxRuntimeError({
             message: 'Cannot process an action without a type.',
-            hint: DetoxInternalError.reportIssue,
+            hint: DetoxRuntimeError.reportIssue,
             debugInfo: action,
           });
         }
