+++ detox/src/server/DetoxServer.js
@@ -1,7 +1,6 @@
 const _ = require('lodash');
 const WebSocket = require('ws');
 const DetoxSessionManager = require('./DetoxSessionManager');
-const DetoxInternalError = require('../errors/DetoxInternalError');
 const DetoxRuntimeError = require('../errors/DetoxRuntimeError');
 const log = require('../utils/logger').child({ __filename });
 const Deferred = require('../utils/Deferred');
@@ -41,7 +40,7 @@ class DetoxServer {
       log.warn({ event: 'ERROR' },
         `Detox server has been closed abruptly! See the error details below:\n`
         + DetoxRuntimeError.format(e) + '\n'
-        + DetoxInternalError.reportIssue
+        + DetoxRuntimeError.reportIssue
       );
     }
   }
